requirements.txt
PyGithub==2.1.1
openai==1.12.0 